Ensure that all uses of <ctype.h> functions are applied to unsigned-char values, whether the local char type is signed or not.  This is necessary for portability.  Per discussion on pghackers around 9/16/00.

+++ src/backend/port/strtoul.c
@@ -96,9 +96,9 @@ register int base;
 			c -= isupper(c) ? 'A' - 10 : 'a' - 10;
 		else
 			break;
-		if (c >= base)
+		if ((int) c >= base)
 			break;
-		if (any < 0 || acc > cutoff || (acc == cutoff && c > cutlim))
+		if (any < 0 || acc > cutoff || (acc == cutoff && (int) c > cutlim))
 			any = -1;
 		else
 		{
